tools/voice-checker/pyproject.toml
[project]
name = "voice-checker"
version = "0.1.0"
description = "Audio quality binary classifier using CNN on mel spectrogram"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.10",
    "torchaudio>=2.10",
    "torchcodec",
    "numpy>=1.26,<2.0",
    "scikit-learn",
    "loguru",
    "pyyaml",
    "tqdm",
    "fastapi[standard]>=0.115.2",
    "uvicorn[standard]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }
